SMS/backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.0,<5.0
Pillow>=10.0,<12.0
PyMySQL>=1.1,<2.0
gunicorn>=22.0,<24.0
whitenoise>=6.7,<7.0
dj-database-url>=2.2,<3.0
psycopg[binary]>=3.2,<4.0
